% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_users_app.R
\name{calculate_connect_daily_user_counts}
\alias{calculate_connect_daily_user_counts}
\title{Process the connect_users data to compute daily metrics used for historical trends}
\usage{
calculate_connect_daily_user_counts(data)
}
\arguments{
\item{data}{A data frame or tibble containing the raw connect_users data.}
}
\description{
Process the connect_users data to compute daily metrics used for historical trends
}
