% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbench_users_app.R
\name{calculate_workbench_daily_user_counts}
\alias{calculate_workbench_daily_user_counts}
\title{Process the pwb_users data to compute daily metrics used for historical trends}
\usage{
calculate_workbench_daily_user_counts(data)
}
\arguments{
\item{data}{A data frame or tibble containing the raw pwb_users data.}
}
\description{
Process the pwb_users data to compute daily metrics used for historical trends
}
