% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chr_data}
\alias{chr_data}
\alias{chr_get_curated_metric_data}
\title{Load Chronicle data}
\usage{
chr_data(metric, base_path)

chr_get_curated_metric_data(metric, base_path)
}
\arguments{
\item{metric}{Name of the curated metric to retrieve (e.g., "connect/user_totals")}

\item{base_path}{Base path to Chronicle data directory}
}
\value{
Arrow dataset object
}
\description{
Loads pre-aggregated (curated) Chronicle metric data. This is the
recommended way to access Chronicle data for most use cases.
}
\details{
For raw/unaggregated data, see \code{\link[=chr_raw_data]{chr_raw_data()}}.
}
\examples{
\dontrun{
# Load curated Connect user totals
data <- chr_data("connect/user_totals", "/var/lib/posit-chronicle/data")

# Load from S3
data <- chr_data("connect/user_totals", "s3://chronicle-bucket/data")
}
}
