% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chr_get_curated_metric_data}
\alias{chr_get_curated_metric_data}
\title{Get curated Chronicle metric data from parquet files}
\usage{
chr_get_curated_metric_data(metric, base_path)
}
\arguments{
\item{metric}{Name of the curated metric to retrieve}

\item{base_path}{Base path to Chronicle data directory}
}
\value{
Arrow dataset object
}
\description{
Get curated Chronicle metric data from parquet files
}
