% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chr_get_metric_data}
\alias{chr_get_metric_data}
\title{Get Chronicle metric data from parquet files}
\usage{
chr_get_metric_data(
  metric,
  base_path,
  frequency = c("daily", "hourly"),
  ymd = NULL,
  schema = NULL
)
}
\arguments{
\item{metric}{Name of the metric to retrieve}

\item{base_path}{Base path to Chronicle data directory}

\item{frequency}{Frequency of data collection ("daily" or "hourly")}

\item{ymd}{Optional list with year, month, day for specific date filtering}

\item{schema}{Optional Arrow schema for the dataset}
}
\value{
Arrow dataset object
}
\description{
Get Chronicle metric data from parquet files
}
