% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chr_path}
\alias{chr_path}
\title{Build path to Chronicle metric data}
\usage{
chr_path(base_path, metric = NULL, frequency = c("daily", "hourly"))
}
\arguments{
\item{base_path}{Base path to Chronicle data directory}

\item{metric}{Name of the metric (e.g., "connect_users")}

\item{frequency}{Frequency of data collection ("daily" or "hourly")}
}
\value{
Character string with the full path to the metric data
}
\description{
Build path to Chronicle metric data
}
\keyword{internal}
