% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle-reports-package.R
\docType{package}
\name{chronicle.reports-package}
\alias{chronicle.reports}
\alias{chronicle.reports-package}
\title{chronicle.reports: Posit Chronicle Reports}
\description{
Run reports for Posit Chronicle. You provide the location of your Chronicle data and this package will generate reports to help you analyze and visualize your data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/posit-dev/chronicle-reports}
  \item Report bugs at \url{https://github.com/posit-dev/chronicle-reports/issues}
}

}
\author{
\strong{Maintainer}: Mark Tucker \email{mark.tucker@posit.co}

Authors:
\itemize{
  \item Matt Urbina \email{matt.urbina@posit.co}
  \item Tim Margheim \email{tim.margheim@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
