% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{chronicle_run_app}
\alias{chronicle_run_app}
\title{Run a Chronicle Reports Shiny App}
\usage{
chronicle_run_app(
  app_name,
  base_path = Sys.getenv("CHRONICLE_BASE_PATH", APP_CONFIG$DEFAULT_BASE_PATH)
)
}
\arguments{
\item{app_name}{Name of the app to run. Available apps:
\itemize{
\item \code{"connect"} - Comprehensive Connect Dashboard
\item \code{"workbench"} - Comprehensive Workbench Dashboard
}
Use \code{\link[=chronicle_list_apps]{chronicle_list_apps()}} to see all available apps.}

\item{base_path}{The base path where Chronicle data files are stored.
Defaults to the value of the \code{CHRONICLE_BASE_PATH} environment variable,
or \code{"/var/lib/posit-chronicle/data"} if the environment variable is not set.
This path should be accessible by the Shiny app.
If you deploy this app to Posit Connect, you can set this environment variable
in the Connect UI. See
\url{https://docs.posit.co/connect/user/content-settings/#content-vars}
for more information.}
}
\value{
This function does not return a value. It launches the Shiny app.
}
\description{
Launch one of the available Chronicle Reports Shiny dashboards. Apps are
located in the package's \verb{inst/apps/} directory and can be listed with
\code{\link[=chronicle_list_apps]{chronicle_list_apps()}}.
}
\examples{
if (interactive()) {
  # List available apps
  chronicle_list_apps()

  # Run an app with default data path
  chronicle_run_app("connect")

  # Run an app with custom filesystem path
  chronicle_run_app("connect", "/path/to/chronicle/data")

  # Run an app with S3 path
  chronicle_run_app("connect", "s3://chronicle-bucket/data")
}
}
