% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_users_app.R
\name{connect_users_app}
\alias{connect_users_app}
\title{Run the Connect Users Dashboard Shiny App}
\usage{
connect_users_app()
}
\description{
Run the Connect Users Dashboard Shiny App
}
\examples{
connect_users_app()
}
