% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chronicle_data}
\alias{chronicle_data}
\title{Load Chronicle data}
\usage{
chronicle_data(metric, base_path)
}
\arguments{
\item{metric}{Name of the curated metric to retrieve (e.g., "connect/user_totals")}

\item{base_path}{Base path to Chronicle data directory}
}
\value{
Arrow dataset object
}
\description{
Loads curated Chronicle metric data. This is the
recommended way to access Chronicle data for most use cases.
}
\examples{
\dontrun{
# Load curated Connect user totals
data <- chronicle_data("connect/user_totals", "/var/lib/posit-chronicle/data")

# Load from S3
data <- chronicle_data("connect/user_totals", "s3://chronicle-bucket/data")
}
}
